Move externals to related modules
