+++ firebase-app/src/jsMain/kotlin/dev/gitlive/firebase/firebase.kt
@@ -4,12 +4,12 @@
 
 package dev.gitlive.firebase
 
-import dev.gitlive.firebase.externals.app.deleteApp
-import dev.gitlive.firebase.externals.app.getApp
-import dev.gitlive.firebase.externals.app.getApps
-import dev.gitlive.firebase.externals.app.initializeApp
+import dev.gitlive.firebase.externals.deleteApp
+import dev.gitlive.firebase.externals.getApp
+import dev.gitlive.firebase.externals.getApps
+import dev.gitlive.firebase.externals.initializeApp
 import kotlin.js.json
-import dev.gitlive.firebase.externals.app.FirebaseApp as JsFirebaseApp
+import dev.gitlive.firebase.externals.FirebaseApp as JsFirebaseApp
 
 actual val Firebase.app: FirebaseApp
     get() = FirebaseApp(getApp())
