+++ firebase-config/src/jsMain/kotlin/dev/gitlive/firebase/remoteconfig/FirebaseRemoteConfig.kt
@@ -3,7 +3,7 @@ package dev.gitlive.firebase.remoteconfig
 import dev.gitlive.firebase.Firebase
 import dev.gitlive.firebase.FirebaseApp
 import dev.gitlive.firebase.FirebaseException
-import dev.gitlive.firebase.externals.remoteconfig.*
+import dev.gitlive.firebase.remoteconfig.externals.*
 import kotlinx.coroutines.await
 import kotlin.js.json
