+++ firebase-database/src/jsMain/kotlin/dev/gitlive/firebase/database/externals/callbacks.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2020 GitLive Ltd.  Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.gitlive.firebase.externals.database
+package dev.gitlive.firebase.database.externals
 
 typealias ChangeSnapshotCallback = (data: DataSnapshot, previousChildName: String?) -> Unit
 typealias ValueSnapshotCallback = (data: DataSnapshot) -> Unit
