+++ firebase-functions/src/jsMain/kotlin/dev/gitlive/firebase/functions/externals/functions.kt
@@ -1,9 +1,9 @@
 @file:JsModule("firebase/functions")
 @file:JsNonModule
 
-package dev.gitlive.firebase.externals.functions
+package dev.gitlive.firebase.functions.externals
 
-import dev.gitlive.firebase.externals.app.FirebaseApp
+import dev.gitlive.firebase.externals.FirebaseApp
 import kotlin.js.Json
 
 external fun connectFunctionsEmulator(functions: Functions, host: String, port: Int)
