+++ firebase-perf/src/jsMain/kotlin/dev/gitlive/firebase/perf/externals/performance.kt
@@ -1,20 +1,20 @@
 @file:JsModule("firebase/performance")
 @file:JsNonModule
 
-package dev.gitlive.firebase.perf
+package dev.gitlive.firebase.perf.externals
 
-import dev.gitlive.firebase.externals.app.FirebaseApp
+import dev.gitlive.firebase.externals.FirebaseApp
 
-external fun getPerformance(app: FirebaseApp? = definedExternally): JsFirebasePerformance
+external fun getPerformance(app: FirebaseApp? = definedExternally): FirebasePerformance
 
-external fun trace(performance: JsFirebasePerformance, name: String): JsPerformanceTrace
+external fun trace(performance: FirebasePerformance, name: String): PerformanceTrace
 
-external interface JsFirebasePerformance {
+external interface FirebasePerformance {
     var dataCollectionEnabled: Boolean
     var instrumentationEnabled: Boolean
 }
 
-external interface JsPerformanceTrace {
+external interface PerformanceTrace {
     fun getAttribute(attr: String): String?
     fun getAttributes(): Map<String, String>
     fun getMetric(metricName: String): Int
