+++ firebase-perf/src/jsMain/kotlin/dev/gitlive/firebase/perf/metrics/Trace.kt
@@ -1,10 +1,10 @@
 package dev.gitlive.firebase.perf.metrics
 
-import dev.gitlive.firebase.perf.JsPerformanceTrace
+import dev.gitlive.firebase.perf.externals.PerformanceTrace
 import dev.gitlive.firebase.perf.rethrow
 
 
-actual class Trace internal constructor(private val js: JsPerformanceTrace) {
+actual class Trace internal constructor(private val js: PerformanceTrace) {
 
     actual fun start() = rethrow { js.start() }
     actual fun stop() = rethrow { js.stop() }
